Remove `Cardano.Node.ConfigCLI` and move  tracer related types to `Cardano.Node.Tracers`
