+++ cardano-node/src/Cardano/Chairman.hs
@@ -61,7 +61,7 @@ import           Ouroboros.Network.Protocol.ChainSync.Codec
 import           Ouroboros.Network.Protocol.Handshake.Version
 import           Ouroboros.Network.NodeToClient
 
-import           Cardano.Node.CLI (TraceConstraints)
+import           Cardano.Node.Tracers (TraceConstraints)
 
 -- | Run chairman: connect with all the core nodes.  Chairman will store the
 -- forks from a common prefix.  If any of them is longer than the security
