+++ cardano-node/src/Cardano/Node/Configuration/Types.hs
@@ -32,6 +32,7 @@ module Cardano.Node.Configuration.Types
 
 import           Cardano.Prelude
 
+
 --------------------------------------------------------------------------------
 -- Cardano Environment
 --------------------------------------------------------------------------------
