+++ cardano-node/src/Cardano/Wallet/Client.hs
@@ -47,7 +47,7 @@ import           Ouroboros.Network.Protocol.ChainSync.Codec
 import           Ouroboros.Network.Protocol.Handshake.Version
 import           Ouroboros.Network.NodeToClient
 
-import           Cardano.Node.CLI (TraceConstraints)
+import           Cardano.Node.Tracers (TraceConstraints)
 
 runWalletClient :: forall blk.
                    ( RunNode blk
